+++ ICD.Connect.Calendaring.Google/GoogleCalendarDevice.cs
@@ -300,16 +300,17 @@ namespace ICD.Connect.Calendaring.Google
 			const string url = "https://www.googleapis.com/oauth2/v4/token";
 
 			//Dispatch!
-			string result;
-			if (!m_Port.Post(url, headers, bodyData, out result))
+			WebPortResponse output = m_Port.Post(url, headers, bodyData);
+
+			if (!output.Success)
 			{
-				Log(eSeverity.Error, "Failed to get token - {0}", result);
+				Log(eSeverity.Error, "Failed to get token - {0}", output.DataAsString);
 				m_Token = null;
 				return null;
 			}
 
 			//Get the token string value out of the JSON
-			GoogleTokenResponse response = JsonConvert.DeserializeObject<GoogleTokenResponse>(result);
+			GoogleTokenResponse response = JsonConvert.DeserializeObject<GoogleTokenResponse>(output.DataAsString);
 			m_Token = response.AccessToken;
 			m_TokenExpireTime = IcdEnvironment.GetLocalTime() + new TimeSpan(0, 0, response.ExpiresInSeconds);
 			return m_Token;
@@ -323,13 +324,13 @@ namespace ICD.Connect.Calendaring.Google
 
 			string url = string.Format("https://www.googleapis.com/calendar/v3/calendars/{0}/events?access_token={1}",
 			                           CalendarId, m_Token);
-			string result;
-			bool success = m_Port.Get(url, out result);
+			
+			WebPortResponse getResponse = m_Port.Get(url);
 
 			GoogleCalendarViewResponse response = null;
-			if (!string.IsNullOrEmpty(result))
-				response = JsonConvert.DeserializeObject<GoogleCalendarViewResponse>(result);
-			if (!success)
+			if (!string.IsNullOrEmpty(getResponse.DataAsString))
+				response = JsonConvert.DeserializeObject<GoogleCalendarViewResponse>(getResponse.DataAsString);
+			if (!getResponse.Success)
 			{
 				if (response != null && response.Error != null)
 					throw new InvalidOperationException(response.Error.Message);